Fix a few new linting failures.
When updating transient packages some new linting failures cropped up
(legitmately). Since we publish the `tests/helpers/` files publicly
there were a few packages that are only used in test helpers that were
not being published.

The fix here is to ignore a few (`which` and `websocket`), but the
correct solution is to move the helpers that we _do_ expect folks to use
publicly (e.g. the `tests/helpers/mock-*.js` ones, or
`tests/helpers/fixturify-project.js`) to a `lib/test-helpers` folder.