Ensure `CLI._environment` is fully resolved.
`CLI` was saving off the passed in argument as `this._environment`,
however that argument is _not_ resolved (it may contain properties that
are lazily resolved promises). When `CLI.prototype.maybeMakeCommand`
uses the stashed off `this._environment` in order to build the command
instance to be used it does not wait for any async.

At the moment, this is "fine" because the specific properties used
inside `maybeMakeCommand` (`this._environment.commands`,
`this._environment.tasks`, `this._environment.project`, and
`this._environment.settings`) are currently synchronously defined, but
a future refactor could **easily** make one of those async and therefore
break the functionality of `maybeMakeCommand`.

This changes the implementation of `CLI.prototype.run` to only set
`this._environment` to the _resolved_ value, avoiding future issues.