Ensure that the captured exit is released.
Prior to this, an error thrown during the bulk of `lib/cli/index.js`
would have caused the process to be captured (which happens very early)
but not released. This is because prior to this change we only ensured
cleanup around the `cli.run` asynchrony, but not for any of the other
work there.