Add config/ember-cli-update.json to app and addon blueprints.
Enables apps and addons to update ember-cli without impacting future
`ember-cli-update` runs.

Without this `ember-cli-update` will use the
current version of `ember-cli` in the `package.json` to determine which
version to attempt to update _from_. Unfortunately, this means that if
you were to update your `ember-cli` version without _also_ running the
full blueprint update, you would have a much harder time of updating
your general project structure in the future.

Support for this format and process was nailed down and merged in
[emberjs/rfcs#477](https://github.com/emberjs/rfcs/blob/master/text/0477-blueprints-update.md).