Remove "process already captured" errors.
This has been a very annoying source of CI failures. If it were
technically more correct I'd say we should keep it around and attempt
to fix the underlying issue, but I do not think it is. In most cases,
we are either setting up on the same object (in which case no need to
release any existing handlers) or we are providing a custom
`MockProcess` object which really won't ever _need_ capturing.