use travis build matrix/expansion instead of include
I learned more about Travis CI stages and wanted to share. Travis has a default stage called "test". This is were all build matrix expansion jobs go. If we move all the `include` env jobs to the real env section, we can allow Travis to expand them into the "test" stage. You can see how it looks [here](https://github.com/kellyselden/ember-a11y-blueprint/pull/58).

Even though this is technically more "correct", it's a little more verbose, and you are stuck with the name "test" instead of how we preferred to name the stages. I'll leave this here for debate on whether it's a good change or not.