Revert testing infrastructure updates (temp fix for 3.23.0).
This reverts the testing infrastructure updates that have been in the
beta cycle since 3.23.0-beta.1 in order to ship ember-cli@3.23.0 (we are
already weeks late). At the moment, the current implementation places
`qunit` and `es6-promise` into the `assets/vendor.js` (on accident). A
number of fixes are on the way but we need to release 3.23.0 before they
are ready for adoption.