Revert "[BUGFIX beta] Comment out IE11 for new apps to avoid deprecation"
This reverts commit 4373c5c3fb623cb6c4eaac7f21710c2ecc3945c8.

The commit was accidentally pushed to the beta branch directly, and
the CI run failed on Node 10 and 12. The failure seems unrelated
though.

Opening this PR to run CI without the commit, if this passes feel
free to land the revert and we can try the fix again through proper
PR process.