Improve dx when WelcomePage is present
This moves the `{{outlet}}` to before the `<WelcomePage/>`. In the out-of-the-box case, it doesn't look any different because there's no content in the index route. But if you try to put content in the index route without first editing application.hbs, this ensures your content will be on screen, above the WelcomePage, rather than hidden off the bottom scroll, which can result in confusion and make you think your change had no effect.