Add support for the `@ember-data/*` packages in the update-blueprint-dependencies.js script
EmberData is no longer in a lockstep version cadence with the ember-source and ember-cli projects. As a result, it's possible that no new ember-data stable release is out when ember-cli bumps the stable deps. This means that the `@ember-data/*` package.json files of the blueprints are published with their `-beta.` versions.

This updates the script so that it uses the same version that we use for the main `ember-data` package.